Fix driftbill
Bill drifting is a glitch that allows a player to drift in bullet bill state, allowing them to hit players by drifting on their face, or landing out-of-bounds when the bullet ends. This is an unfair behavior for other players as it is very hard to dodge. This glitch can be triggered as long as the jumped attribute of the player is set to true when the player enters in bullet bill state. Removing this attribute upon using the item patches the issue (showcased by Kirio here: https://discord.com/channels/308979137480097793/1300233719222767737/1301881702381260831)