Various fix (#61)
* Fix misaligned sprites (#59)

* Fix misaligned sprites

Made the Chomp's eyes black instead of transparent

* Fix Chomp

---------

Co-authored-by: BuildTools <unconfigured@null.spigotmc.org>

* Fix driftbill

Bill drifting is a glitch that allows a player to drift in bullet bill state, allowing them to hit players by drifting on their face, or landing out-of-bounds when the bullet ends. This is an unfair behavior for other players as it is very hard to dodge. This glitch can be triggered as long as the jumped attribute of the player is set to true when the player enters in bullet bill state. Removing this attribute upon using the item patches the issue (showcased by Kirio here: https://discord.com/channels/308979137480097793/1300233719222767737/1301881702381260831)

* Make flows affect players in mega / star / spinning state

Requested: https://discord.com/channels/308979137480097793/426051260467249168/1301677375561404449

* Prioritize boost ramps overr jump ramps

This fix a glitch were you often skip boost ramps before jump ramps, it happens a lot in 200cc

---------

Co-authored-by: BuildTools <unconfigured@null.spigotmc.org>