Fix Boo Lake CPU glitch
Move the respawn point of the hole after the first jump on the next platform. If a CPU were to be target shocked above this hole, they would respawn before the jump and use it to fall off on the left for some reason. The change here is this hole data, old: [164,0,88,48,143,26], new: [164,0,88,48,260,26]